Added help.R file
Added a help.R file which should contain the helper functions and sourced (source(data=x)) it in the project.Rmd file. Sourcing instead of loading the chunk via knitr since we don't need to call it in the Rmd file.

Since much of data_abbreviations was redundant in regard to the data_readme I saved everthing that was not in the data_readme at the end of said data_readme and deleted the data_abbrevations file.